requirements.txt
pytesseract
Pillow
opencv-python
numpy
transformers
torch
fairseq
indic-nlp-library
flask
streamlit
pdf2image
reportlab
matplotlib
spellchecker
fuzzywuzzy
python-Levenshtein
google-generativeai
googletrans==4.0.0-rc1
tensorflow
fastapi[all]
tf-keras
hf_xet
